pyproject.toml
[tool.poetry]
name = "dnd-characters"
version = "0.1.0"
description = ""
authors = ["Rinat Akhmetov <appvshare@gmail.com>"]
readme = "README.md"
packages = [{include = "dnd_characters"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.31.1"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
